move 'naming.js' to 'shared' directory
