+++ lib/cli-engine/config-array-factory.js
@@ -38,9 +38,9 @@ const path = require("path");
 const importFresh = require("import-fresh");
 const stripComments = require("strip-json-comments");
 const { validateConfigSchema } = require("../shared/config-validator");
+const naming = require("../shared/naming");
 const ModuleResolver = require("../shared/relative-module-resolver");
 const { ConfigArray, ConfigDependency, OverrideTester } = require("./config-array");
-const naming = require("./naming");
 const debug = require("debug")("eslint:config-array-factory");
 
 //------------------------------------------------------------------------------
