+++ lib/shared/naming.js
